+++ src/sentry/integrations/gitlab/issue_sync.py
@@ -1,20 +1,16 @@
 from __future__ import annotations
 
 import logging
-from collections.abc import Mapping, MutableMapping
+from collections.abc import Mapping
 from typing import Any
 from urllib.parse import quote
-
-from django.utils.translation import gettext_lazy as _
 
 from sentry import features
 from sentry.integrations.mixins.issues import IssueSyncIntegration, ResolveSyncAction
 from sentry.integrations.models.external_actor import ExternalActor
 from sentry.integrations.models.external_issue import ExternalIssue
 from sentry.integrations.models.integration_external_project import IntegrationExternalProject
-from sentry.integrations.services.integration import integration_service
 from sentry.integrations.types import EXTERNAL_PROVIDERS_REVERSE, ExternalProviderEnum
-from sentry.organizations.services.organization import organization_service
 from sentry.shared_integrations.exceptions import ApiError
 from sentry.users.services.user.model import RpcUser
 from sentry.users.services.user.service import user_service
@@ -40,13 +36,13 @@ class GitlabIssueSyncSpec(IssueSyncIntegration):
     ) -> tuple[str, str] | tuple[None, None]:
         """
         Split the external issue key into project path and issue iid.
-        Format is "{domain_name}:{path_with_namespace}#{issue_iid}"
+        Format is "{domain_name}:{path_with_namespace}#{issue_id}"
         """
         try:
-            _, project_and_issue_iid = external_issue_key.split(":", 1)
-            project_id, issue_iid = project_and_issue_iid.split("#")
+            _, project_and_issue_id = external_issue_key.split(":", 1)
+            project_id, issue_id = project_and_issue_id.split("#")
 
-            return project_id, issue_iid
+            return project_id, issue_id
         except ValueError:
             logger.exception(
                 "split-external-key.invalid-format",
@@ -70,9 +66,9 @@ class GitlabIssueSyncSpec(IssueSyncIntegration):
         """
         client = self.get_client()
 
-        project_id, issue_iid = self.split_external_issue_key(external_issue.key)
+        project_id, issue_id = self.split_external_issue_key(external_issue.key)
 
-        if not project_id or not issue_iid:
+        if not project_id or not issue_id:
             logger.error(
                 "assignee-outbound.invalid-key",
                 extra={
@@ -151,10 +147,11 @@ class GitlabIssueSyncSpec(IssueSyncIntegration):
 
         # Update GitLab issue assignees
         try:
-            data = {"assignee_ids": [gitlab_user_id] if gitlab_user_id else []}
             # URL-encode project_id since it's a path_with_namespace (e.g., "owner/repo")
             encoded_project_id = quote(project_id, safe="")
-            client.update_issue_assignees(encoded_project_id, issue_iid, data)
+            client.update_issue_assignees(
+                encoded_project_id, issue_id, [gitlab_user_id] if gitlab_user_id else []
+            )
         except Exception as e:
             self.raise_error(e)
 
@@ -189,84 +186,6 @@ class GitlabIssueSyncSpec(IssueSyncIntegration):
         config["sync_status_forward"] = sync_status_forward
         return config
 
-    def _get_organization_config_default_values(self) -> list[dict[str, Any]]:
-        """
-        Return configuration options for the GitLab integration.
-        """
-        config: list[dict[str, Any]] = []
-
-        if self.check_feature_flag():
-            config.extend(
-                [
-                    {
-                        "name": self.inbound_assignee_key,
-                        "type": "boolean",
-                        "label": _("Sync GitLab Assignment to Sentry"),
-                        "help": _(
-                            "When an issue is assigned in GitLab, assign its linked Sentry issue to the same user."
-                        ),
-                        "default": False,
-                    },
-                    {
-                        "name": self.outbound_assignee_key,
-                        "type": "boolean",
-                        "label": _("Sync Sentry Assignment to GitLab"),
-                        "help": _(
-                            "When an issue is assigned in Sentry, assign its linked GitLab issue to the same user."
-                        ),
-                        "default": False,
-                    },
-                    {
-                        "name": self.comment_key,
-                        "type": "boolean",
-                        "label": _("Sync Sentry Comments to GitLab"),
-                        "help": _("Post comments from Sentry issues to linked GitLab issues"),
-                    },
-                ]
-            )
-
-        return config
-
-    def get_organization_config(self) -> list[dict[str, Any]]:
-        """
-        Return configuration options for the GitLab integration.
-        """
-        config = self._get_organization_config_default_values()
-
-        context = organization_service.get_organization_by_id(
-            id=self.organization_id, include_projects=False, include_teams=False
-        )
-        assert context, "organizationcontext must exist to get org"
-        organization = context.organization
-
-        has_issue_sync = features.has("organizations:integrations-issue-sync", organization)
-
-        if not has_issue_sync:
-            for field in config:
-                field["disabled"] = True
-                field["disabledReason"] = _(
-                    "Your organization does not have access to this feature"
-                )
-
-        return config
-
-    def update_organization_config(self, data: MutableMapping[str, Any]) -> None:
-        """
-        Update the configuration field for an organization integration.
-        """
-        if not self.org_integration:
-            return
-
-        config = self.org_integration.config
-
-        config.update(data)
-        org_integration = integration_service.update_organization_integration(
-            org_integration_id=self.org_integration.id,
-            config=config,
-        )
-        if org_integration is not None:
-            self.org_integration = org_integration
-
     def create_comment_attribution(self, user_id, comment_text):
         user = user_service.get_user(user_id)
         username = "Unknown User" if user is None else user.name
@@ -278,14 +197,14 @@ class GitlabIssueSyncSpec(IssueSyncIntegration):
     def update_comment(self, issue_id, user_id, group_note):
         quoted_comment = self.create_comment_attribution(user_id, group_note.data["text"])
 
-        project_id, issue_iid = self.split_external_issue_key(issue_id)
+        project_id, parsed_issue_id = self.split_external_issue_key(issue_id)
 
-        if not project_id or not issue_iid:
+        if not project_id or not parsed_issue_id:
             logger.error(
                 "update-comment.invalid-key",
                 extra={
                     "provider": self.model.provider,
-                    "external_issue_key": issue_id,
+                    "external_issue_key": parsed_issue_id,
                 },
             )
             return None
@@ -294,7 +213,7 @@ class GitlabIssueSyncSpec(IssueSyncIntegration):
 
         return self.get_client().update_comment(
             encoded_project_id,
-            issue_iid,
+            parsed_issue_id,
             group_note.data["external_id"],
             {"body": quoted_comment},
         )
@@ -303,14 +222,14 @@ class GitlabIssueSyncSpec(IssueSyncIntegration):
         comment = group_note.data["text"]
         quoted_comment = self.create_comment_attribution(user_id, comment)
 
-        project_id, issue_iid = self.split_external_issue_key(issue_id)
+        project_id, parsed_issue_id = self.split_external_issue_key(issue_id)
 
-        if not project_id or not issue_iid:
+        if not project_id or not parsed_issue_id:
             logger.error(
                 "create-comment.invalid-key",
                 extra={
                     "provider": self.model.provider,
-                    "external_issue_key": issue_id,
+                    "external_issue_key": parsed_issue_id,
                 },
             )
             return None
@@ -318,5 +237,5 @@ class GitlabIssueSyncSpec(IssueSyncIntegration):
         encoded_project_id = quote(project_id, safe="")
 
         return self.get_client().create_comment(
-            encoded_project_id, issue_iid, {"body": quoted_comment}
+            encoded_project_id, parsed_issue_id, {"body": quoted_comment}
         )
