+++ tests/sentry/integrations/gitlab/test_webhook.py
@@ -72,7 +72,7 @@ class WebhookTest(GitLabTestCase):
         assert response.status_code == 400
         assert (
             response.reason_phrase
-            == "The customer has edited the webhook in Gitlab to include other types of events. We only support these kinds of events: Merge Request Hook, Push Hook"
+            == "The customer has edited the webhook in Gitlab to include other types of events. We only support these kinds of events: Issue Hook, Merge Request Hook, Push Hook"
         )
 
     def test_invalid_token(self) -> None:
